xong luồng chuẩn đặt món và bếp thao tác

+++ aspnet-core/src/SmartRestaurant.Application.Contracts/InventoryManagement/Ingredients/Dto/CreateUpdateIngredientDto.cs
@@ -19,7 +19,7 @@ public class CreateUpdateIngredientDto
     [Required]
     public Guid UnitId { get; set; }
     
-    public decimal? CostPerUnit { get; set; }
+    public int? CostPerUnit { get; set; }
     
     [MaxLength(512)]
     public string? SupplierInfo { get; set; }
