+++ aspnet-core/src/SmartRestaurant.Application.Contracts/InventoryManagement/Ingredients/Dto/IngredientPurchaseUnitDto.cs
@@ -10,6 +10,6 @@ public class IngredientPurchaseUnitDto : FullAuditedEntityDto<Guid>
     public string UnitName { get; set; } = string.Empty;
     public int ConversionRatio { get; set; }
     public bool IsBaseUnit { get; set; }
-    public decimal? PurchasePrice { get; set; }
+    public int? PurchasePrice { get; set; }
     public bool IsActive { get; set; }
 }
